move everything to the home page
