+++ src/pages/index.js
@@ -1,12 +1,15 @@
 /** @jsx jsx */
 import { jsx } from 'theme-ui';
 import { Helmet } from 'react-helmet';
-import { Link } from 'gatsby';
-import { StaticImage } from 'gatsby-plugin-image';
-import Layout from '../components/layout';
 import { BsArrowRight } from 'react-icons/bs';
+import { Link } from 'gatsby';
+
+import Layout from '../components/layout';
 import { ProjectCard } from '../components/projectCard';
-import { TrybeLink } from '../components/trybeLink';
+import CoverSection from '../components/coverSection';
+import ContactSection from '../components/contactSection';
+import AboutSection from '../components/aboutSection';
+import SkillsSection from '../components/skillsSection';
 
 const IndexPage = () => {
   return (
@@ -14,31 +17,11 @@ const IndexPage = () => {
       <Helmet>
         <title>Matheus Gaspar - web developer</title>
       </Helmet>
-      <MainHeading
-        heading="Matheus Gaspar"
-        subheading="Full-stack web developer"
-        sx={{
-          my: 5,
-        }}
-      />
 
-      <section sx={{ mb: 4 }}>
-        <p>
-          I'm a 29-year-old engineer from Brazil, taking the first steps in my
-          career as a <strong>software developer</strong>. Currently, I'm
-          studying front and back-end web development at the awesome school{' '}
-          <TrybeLink />, with focus on <strong>JavaScript</strong>,{' '}
-          <strong>React</strong> and <strong>Node.js</strong>.
-        </p>
-        <p>
-          This is my personal spot in the internet - you can learn more{' '}
-          <Link to="/about">about me</Link> or check what I've been{' '}
-          <Link to="/projects">working on</Link> recently.
-        </p>
-      </section>
+      <CoverSection />
 
       <section>
-        <h2>Latest projects</h2>
+        <h2>Projects</h2>
         <div
           sx={{
             display: 'flex',
@@ -79,52 +62,11 @@ const IndexPage = () => {
           </Link>
         </div>
       </section>
+      <AboutSection />
+      <SkillsSection />
+      <ContactSection />
     </Layout>
   );
 };
-
-const MainHeading = ({ heading, subheading, ...props }) => (
-  <header
-    {...props}
-    sx={{
-      display: 'flex',
-      justifyContent: 'space-between',
-      alignItems: 'flex-start',
-      pr: 5,
-    }}
-  >
-    <div>
-      <h1
-        sx={{
-          fontFamily: 'heading',
-          fontWeight: 'extra',
-          letterSpacing: -1,
-          fontSize: 7,
-          m: 0,
-        }}
-      >
-        {heading}
-      </h1>
-      <h2
-        sx={{
-          m: 0,
-          fontFamily: 'body',
-          fontWeight: 'light',
-          fontSize: 4,
-        }}
-      >
-        {subheading}
-      </h2>
-    </div>
-    <div>
-      <StaticImage
-        width={160}
-        placeholder="blurred"
-        alt="Matheus Gaspar"
-        src="../images/profile.png"
-      />
-    </div>
-  </header>
-);
 
 export default IndexPage;
